Renames prefixedDatabase pkg to prefixdb
